Add a framework for regression tests.
Regression tests run insane with a set of arguments and compare the
output with a reference. This allows to assess that the output did not
change between the version of insane with which the reference has been
generated and the current version of insane. It does not test that the
result is valid!

Run tests/test_regression.py as a script to generate the reference
files. Run it with pytest or nosetest to execute the tests.

As is, the framework does not work because the output of insane is
randomize. This needs to be fixed.