Script create zip package
We want to make a module of insane that would be split in multiple
files. Already we depend on simopt, so the parsing of the arguments in
factored out of insane.py.

This makes insane more difficult to use for users that are not familiar
with installing python modules using pip.

A solution to this problem is to distribute insane as a zipped package.
That package is a zip file that can be executed by python. The only
constraint is that the zip file must contain a __main__.py, then it can
be called from the command line as `python insane.zip`. Adding a shebang
on top of the file allows to execute directly `./insane.zip`.

This commit introduces `zippackage.sh` that creates such a zip package
for insane. The zip package is simply named `insane` as the zip
extension would be confusing for most users. The script downloads the
latest version of simopt from github and includes it in the package.

Note that `zippackage.sh` *must* be executed from the 'maintainers'
directory!

The package is build in the 'maintainers' directory and can be tested
with `./insane -h`.