Change reference PDB file with new title
The title for systems with a protein in a membrane use to differ between
PDB and GRO files. GRO files would acknowledge the protein in the title
while PDB would not. With the title generation being factored out of the
writing functions, the title is now the same for bot formats; in both
format, the protein is acknowledge.

This commit update a reference file to take the title change into
account.