Make resource files available in the zip package
Non-python files cannot be read in a zip package as they usually are.
Indeed, in a zip package, resource files leave in memory instead of on
the disk. As a result, they do not have a file name and cannot be
opened.

This commit introduces the function `utils.iter_resource` that should be
used to access resource files instead of `open`. It takes the name of the
resource file expressed relative to the module root as an argument and
return an iterator over the lines of the file.