Regression tests work by directories
Initially, regression tests would save only the gro output, the top
output if any, and the standard output and error. They were saved in
files that had the test name (arguments or alias) and an extension.

This commit makes each test work on a directory. Now each test has a
directory as reference so we are sure we are not missing any produced file.

This is mostly to stay in line with what is done with martinize.